-- Update Candidate endpoint added

+++ KHRMS.Core/Interfaces/IGenericRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> GetById(int id);
+        Task<T> GetById(long id);
         Task<IEnumerable<T>> GetAll();
         Task Add(T entity);
         void Delete(T entity);
